+++ tests/nn/data_parallel/test_features_sharded_ddp.py
@@ -86,7 +86,7 @@ def run_one_step(
     for i in range(5):
         _ = optimizer.step(closure=closure)
         # when running on cpu/gloo the "nodes" are not really different
-        same_params = device == torch.device("cpu") or grad_accumulation
+        same_params = device == torch.device("cpu") or not grad_accumulation
         check_same_models_across_ranks(
             ddp_model, dist.group.WORLD, params_should_be_equal=same_params, check_broadcast_buffers=broadcast_buffers
         )
@@ -359,7 +359,7 @@ def run_test_gpt2(rank, world_size, backend, device, temp_file_name):
         embed_dim=256, num_heads=2, num_layers=12, num_positions=INPUT_DIM * INPUT_DIM, num_vocab=512, num_classes=2
     ).to(device)
     optimizer = OSS(params=model.parameters(), optim=torch.optim.SGD, lr=1e-3, momentum=0.99)
-    ddp_model = ShardedDataParallel(model, optimizer, reduce_buffer_size=0)
+    ddp_model = ShardedDataParallel(model, optimizer)
 
     # Optim loop
     def closure():